refactor: AuthenticationViewModel로 통합
- 사용법
```Swift
// 싱글톤으로 선언해 사용
@StateObject var authViewModel = AuthenticationViewModel.shared
```
