+++ TrackUs/TrackUs/Sources/Authentication/View/AgeGenderView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct AgeGenderView: View {
-    @EnvironmentObject var userInfoViewModel : UserInfoViewModel
+    @StateObject var authViewModel = AuthenticationViewModel.shared
     @Binding private var signUpFlow: SignUpFlow
     
     // nickName 데이터값 변경
@@ -43,8 +43,8 @@ struct AgeGenderView: View {
             Spacer()
             
             MainButton(active: selectedAge != nil && selectedGender != nil, buttonText: "다음으로") {
-                userInfoViewModel.userInfo.age = Int(selectedAge!)
-                userInfoViewModel.userInfo.gender = selectedGender
+                authViewModel.userInfo.age = Int(selectedAge!)
+                authViewModel.userInfo.gender = selectedGender
                 signUpFlow = .runningStyle
             }
         }
