+++ TrackUs/TrackUs/Sources/Authentication/View/NickNameView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct NickNameView: View {
     @Binding private var signUpFlow: SignUpFlow
-    @EnvironmentObject var userInfoViewModel : UserInfoViewModel
+    @StateObject var authViewModel = AuthenticationViewModel.shared
     
     // nickName 데이터값 변경
     @State private var nickName: String = ""
@@ -32,7 +32,7 @@ struct NickNameView: View {
             Spacer()
             MainButton(active: availability, buttonText: "다음으로") {
                 signUpFlow = .profile
-                userInfoViewModel.userInfo.username = nickName
+                authViewModel.userInfo.username = nickName
             }
         }
     }
