+++ TrackUs/TrackUs/Sources/Authentication/View/PhysicalView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct PhysicalView: View {
-    @EnvironmentObject var userInfoViewModel : UserInfoViewModel
+    @StateObject var authViewModel = AuthenticationViewModel.shared
     @Binding private var signUpFlow: SignUpFlow
     
     // nickName 데이터값 변경
@@ -30,8 +30,8 @@ struct PhysicalView: View {
             Spacer()
             
             MainButton(active: weight != nil && height != nil, buttonText: "다음으로") {
-                userInfoViewModel.userInfo.weight = Int(weight!)
-                userInfoViewModel.userInfo.height = Int(height!)
+                authViewModel.userInfo.weight = Int(weight!)
+                authViewModel.userInfo.height = Int(height!)
                 signUpFlow = .ageGender
             }
         }
