+++ TrackUs/TrackUs/Sources/Authentication/View/ProfileImageView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct ProfileImageView: View {
-    @EnvironmentObject var userInfoViewModel : UserInfoViewModel
+    @StateObject var authViewModel = AuthenticationViewModel.shared
     @Binding private var signUpFlow: SignUpFlow
     @State private var image: Image?
     @State private var availability: Bool = false
@@ -27,7 +27,7 @@ struct ProfileImageView: View {
             Spacer()
             
             MainButton(active: image != nil, buttonText: "다음으로") {
-                userInfoViewModel.persistImageToStorage(image: image)
+                authViewModel.persistImageToStorage(image: image)
                 signUpFlow = .physical
             }
         }
