+++ TrackUs/TrackUs/Sources/Root/ContentView.swift
@@ -9,30 +9,28 @@ import SwiftUI
 
 struct ContentView: View {
     @EnvironmentObject var router: Router
-    @StateObject private var viewModel = AuthenticationViewModel()
+    @StateObject var authViewModel = AuthenticationViewModel.shared
     
     var body: some View {
         VStack{
-            switch viewModel.authenticationState {
+            switch authViewModel.authenticationState {
                 // 로그인
             case .unauthenticated, .authenticating:
                 VStack {
                     LoginView()
-                        .environmentObject(viewModel)                }
+                }
                 //이전에 로그인 했을경우
             case .signUpcating:
                 VStack{
                     SignUpView()
-                        .environmentObject(viewModel)
                 }
             case .authenticated:
                 VStack {
                     MainTabView()
-                        .environmentObject(viewModel)
                 }
             }
         }
-        .animation(.easeIn(duration: 0.15), value: viewModel.authenticationState)
+        .animation(.easeIn(duration: 0.15), value: authViewModel.authenticationState)
     }
 }
